WIP: Updated symlink & thirdpartyenricher

+++ pkg/utils/events.go
@@ -9,7 +9,7 @@ type K8sEvent interface {
 
 type EnrichEvent interface {
 	GetBaseEvent() *types.Event
-	GetPID() int
+	GetPID() uint64
 	SetExtra(extra interface{})
 }
